fixup a few profile issues
