+++ src/renderer/src/pages/ProfileEditor.tsx
@@ -1,10 +1,10 @@
 import { useCallback, useEffect, useMemo, useReducer, useRef, useState } from "react";
-import { createPortal } from "react-dom";
 import { useNavigate } from "react-router-dom";
 
 import { MinecraftButton, GRAY_MINECRAFT_BUTTON } from "@renderer/components/MinecraftButton";
 import { PopupPanel, usePopupClose } from "@renderer/components/PopupPanel";
 import { TextInput } from "@renderer/components/TextInput";
+import { ThreeDotsMenu } from "@renderer/components/ThreeDotsMenu";
 import { Popup, PopupUseArguments } from "@renderer/states/PopupStore";
 import { useAppStore } from "@renderer/states/AppStore";
 import { VersionPickerPopup, VersionPickerResult } from "@renderer/popups/VersionPickerPopup";
@@ -21,20 +21,11 @@ function getModIconPath(modsPath: string, modName: string): string | null {
     return fs.existsSync(iconPath) ? iconPath : null;
 }
 
-function AddContentPopup({ submit: rawSubmit }: PopupUseArguments<string | "browse" | null>) {
+function AddContentPopup({ submit: rawSubmit }: PopupUseArguments<"browse" | null>) {
     const animateClose = usePopupClose();
-    const submit = (val: string | "browse" | null) => animateClose(() => rawSubmit(val));
-    const mods = useAppStore(state => state.allValidMods);
-    const activeMods = useAppStore(state => state.allProfiles)[useAppStore(state => state.selectedProfile)]?.mods ?? [];
+    const submit = (val: "browse" | null): void => animateClose(() => rawSubmit(val));
     const selectedProfileUuid = useAppStore(state => state.allProfiles[state.selectedProfile]?.uuid ?? "");
     const modsPath = useMemo(() => GetProfileModsPath(selectedProfileUuid), [selectedProfileUuid]);
-    const availableMods = useMemo(() => mods.filter(m => !activeMods.includes(m)), [mods, activeMods]);
-    const [search, setSearch] = useState("");
-    const filtered = useMemo(() => {
-        if (!search) return availableMods;
-        const q = search.toLowerCase();
-        return availableMods.filter(m => m.toLowerCase().includes(q));
-    }, [availableMods, search]);
 
     return (
         <PopupPanel onExit={() => submit(null)}>
@@ -54,85 +45,6 @@ function AddContentPopup({ submit: rawSubmit }: PopupUseArguments<string | "brow
                     </div>
                 </div>
                 <div className="version-picker-divider" />
-                <div style={{ padding: "8px" }}>
-                    <div className="mod-search-box">
-                        <svg
-                            className="mod-search-icon"
-                            width="14"
-                            height="14"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            stroke="#6f6f6f"
-                            strokeWidth="2.5"
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                        >
-                            <circle cx="11" cy="11" r="8" />
-                            <line x1="21" y1="21" x2="16.65" y2="16.65" />
-                        </svg>
-                        <input
-                            type="text"
-                            className="minecraft-seven mod-search-input"
-                            spellCheck={false}
-                            placeholder="Search local mods..."
-                            value={search}
-                            onInput={e => setSearch(e.currentTarget.value)}
-                        />
-                    </div>
-                </div>
-                <div className="version-picker-list scrollbar">
-                    {filtered.length === 0 && (
-                        <p
-                            className="minecraft-seven"
-                            style={{ color: "#9f9f9f", padding: "12px", textAlign: "center" }}
-                        >
-                            {search ? "No mods match your search." : "No local mods available."}
-                        </p>
-                    )}
-                    {filtered.map(mod => {
-                        const iconPath = getModIconPath(modsPath, mod);
-                        return (
-                            <div
-                                key={mod}
-                                className="version-picker-item"
-                                style={{ justifyContent: "flex-start", gap: 10, padding: "4px 6px" }}
-                                onClick={() => submit(mod)}
-                            >
-                                <div className="profile-editor-mod-icon">
-                                    {iconPath ? (
-                                        <img
-                                            src={`file://${iconPath}`}
-                                            width="36"
-                                            height="36"
-                                            className="pixelated"
-                                            style={{ borderRadius: 3 }}
-                                            alt=""
-                                        />
-                                    ) : (
-                                        <svg width="16" height="16" viewBox="0 0 16 16" fill="none">
-                                            <rect
-                                                x="2"
-                                                y="2"
-                                                width="12"
-                                                height="12"
-                                                rx="2"
-                                                stroke="#6f6f6f"
-                                                strokeWidth="1.5"
-                                            />
-                                            <path
-                                                d="M5 8h6M8 5v6"
-                                                stroke="#6f6f6f"
-                                                strokeWidth="1.5"
-                                                strokeLinecap="round"
-                                            />
-                                        </svg>
-                                    )}
-                                </div>
-                                <p className="minecraft-seven">{mod}</p>
-                            </div>
-                        );
-                    })}
-                </div>
                 <div className="version-picker-divider" />
                 <div className="version-picker-footer" style={{ justifyContent: "flex-start", gap: 8 }}>
                     <MinecraftButton
@@ -154,18 +66,13 @@ function AddContentPopup({ submit: rawSubmit }: PopupUseArguments<string | "brow
 
 export function ProfileEditor() {
     const [profileName, setProfileName] = useState("");
-    const [profileActiveMods, setProfileActiveMods] = useState<string[]>([]);
     const [profileRuntime, setProfileRuntime] = useState<string>("");
     const [profileMinecraftVersion, setProfileMinecraftVersion] = useState<string>("");
     const [profileVersionUuid, setProfileVersionUuid] = useState<string | null>(null);
     const [modSearch, setModSearch] = useState("");
-
-    const [showMenu, setShowMenu] = useState(false);
-    const dotsRef = useRef<HTMLDivElement>(null);
-    const dropdownRef = useRef<HTMLDivElement>(null);
-    const [dropdownPos, setDropdownPos] = useState({ top: 0, right: 0 });
     const [, forceUpdate] = useReducer(x => x + 1, 0);
-    const allValidMods = useAppStore(state => state.allValidMods);
+
+    const allMods = useAppStore(state => state.allMods);
     const allProfiles = useAppStore(state => state.allProfiles);
     const setAllProfiles = useAppStore(state => state.setAllProfiles);
     const selectedProfile = useAppStore(state => state.selectedProfile);
@@ -201,31 +108,32 @@ export function ProfileEditor() {
         const profile = allProfiles[selectedProfile];
         setProfileName(profile?.name ?? "New Profile");
         setProfileRuntime(profile?.runtime ?? "Vanilla");
-        setProfileActiveMods(profile?.mods ?? []);
         setProfileMinecraftVersion(profile?.minecraft_version ?? "1.21.0.3");
         setProfileVersionUuid(profile?.version_uuid ?? null);
         profileLoaded.current = true;
+        useAppStore.getState().refreshAllMods();
     }, [allProfiles, selectedProfile]);
 
-    // Auto-save on every change (skip until profile is loaded to avoid overwriting with empty state)
     useEffect(() => {
         if (!profileLoaded.current) return;
         const profile = allProfiles[selectedProfile];
         if (!profile) return;
         profile.name = profileName;
-        profile.runtime = profileActiveMods.length > 0 ? profileRuntime : "Vanilla";
-        profile.mods = profileActiveMods;
+        profile.runtime = profileRuntime;
         profile.minecraft_version = profileMinecraftVersion;
         profile.version_uuid = profileVersionUuid;
         saveData();
-    }, [profileName, profileRuntime, profileActiveMods, profileMinecraftVersion, profileVersionUuid]);
+    }, [profileName, profileRuntime, profileMinecraftVersion, profileVersionUuid]);
+
+    useEffect(() => {
+        loadProfile();
+    }, [loadProfile]);
 
     const removeMod = async (modName: string) => {
         const modPath = path.join(modsPath, modName);
         if (fs.existsSync(modPath)) {
             fs.rmSync(modPath, { recursive: true, force: true });
         }
-        setProfileActiveMods(profileActiveMods.filter(m => m !== modName));
         useAppStore.getState().refreshAllMods();
     };
 
@@ -253,51 +161,15 @@ export function ProfileEditor() {
         if (profile) await doLaunchProfile(profile);
     };
 
-    // Close menu on outside click
-    useEffect(() => {
-        if (!showMenu) return;
-        const handleClick = (e: MouseEvent) => {
-            if (
-                dotsRef.current &&
-                !dotsRef.current.contains(e.target as Node) &&
-                dropdownRef.current &&
-                !dropdownRef.current.contains(e.target as Node)
-            ) {
-                setShowMenu(false);
-            }
-        };
-        document.addEventListener("mousedown", handleClick);
-        return () => document.removeEventListener("mousedown", handleClick);
-    }, [showMenu]);
-
-    useEffect(() => {
-        if (!showMenu || !dotsRef.current) return;
-        const rect = dotsRef.current.getBoundingClientRect();
-        setDropdownPos({
-            top: rect.bottom + 10,
-            right: window.innerWidth - rect.right,
-        });
-    }, [showMenu]);
-
-    useEffect(() => {
-        loadProfile();
-    }, [loadProfile]);
-
     const openAddContent = async () => {
-        const result = await Popup.useAsync<string | "browse" | null>(props => {
+        const result = await Popup.useAsync<"browse" | null>(props => {
             return <AddContentPopup {...props} />;
         });
 
-        if (result === null) return;
-        if (result === "browse") {
-            const profile = allProfiles[selectedProfile];
-            if (profile) useAppStore.getState().setInstallingForProfile(profile.uuid);
-            navigate("/mod-discovery");
-            return;
-        }
-        if (!profileActiveMods.includes(result)) {
-            setProfileActiveMods([...profileActiveMods, result]);
-        }
+        if (result !== "browse") return;
+        const profile = allProfiles[selectedProfile];
+        if (profile) useAppStore.getState().setInstallingForProfile(profile.uuid);
+        navigate("/mod-discovery");
     };
 
     const openVersionPicker = async () => {
@@ -307,6 +179,12 @@ export function ProfileEditor() {
         if (!result) return;
         setProfileMinecraftVersion(result.minecraft_version);
         setProfileVersionUuid(result.version_uuid);
+    };
+
+    const openProfileFolder = () => {
+        const paths = useAppStore.getState().platform.getPaths();
+        const uuid = allProfiles[selectedProfile]?.uuid;
+        if (uuid) shell.openPath(path.join(paths.profilesPath, uuid));
     };
 
     const installedVersions = useMemo(() => versionManager.getInstalledVersions(), [versionManager]);
@@ -320,12 +198,12 @@ export function ProfileEditor() {
     }, [profileVersionUuid, profileMinecraftVersion, installedVersions]);
 
     const allModsList = useMemo(() => {
-        return profileActiveMods.map(name => ({
-            name,
-            exists: allValidMods.includes(name),
-            isDownloading: downloadingMods.includes(name),
+        return allMods.map(mod => ({
+            name: mod.id,
+            exists: mod.ok,
+            isDownloading: downloadingMods.includes(mod.id),
         }));
-    }, [profileActiveMods, allValidMods, downloadingMods]);
+    }, [allMods, downloadingMods]);
 
     const filteredModsList = useMemo(() => {
         if (!modSearch) return allModsList;
@@ -398,47 +276,19 @@ export function ProfileEditor() {
                                     style={{ "--mc-button-container-h": "36px" }}
                                 />
                             </div>
-                            <div className="launcher-profile-card-menu" onClick={e => e.stopPropagation()}>
-                                <div
-                                    className="launcher-profile-card-dots"
-                                    ref={dotsRef}
-                                    onClick={() => setShowMenu(!showMenu)}
-                                >
-                                    <svg width="16" height="16" viewBox="0 0 16 16" fill="none">
-                                        <circle cx="8" cy="3" r="1.5" fill="#FFFFFF" />
-                                        <circle cx="8" cy="8" r="1.5" fill="#FFFFFF" />
-                                        <circle cx="8" cy="13" r="1.5" fill="#FFFFFF" />
-                                    </svg>
-                                </div>
-                                {showMenu &&
-                                    createPortal(
-                                        <div
-                                            className="launcher-profile-card-dropdown"
-                                            ref={dropdownRef}
-                                            style={{ top: dropdownPos.top, right: dropdownPos.right }}
-                                            onClick={e => e.stopPropagation()}
-                                        >
-                                            <div
-                                                className="launcher-profile-card-dropdown-item launcher-profile-card-dropdown-item--danger"
-                                                onClick={() => {
-                                                    deleteProfile();
-                                                    setShowMenu(false);
-                                                }}
-                                            >
-                                                <svg width="14" height="14" viewBox="0 0 16 16" fill="none">
-                                                    <path
-                                                        d="M2 4H14M5.5 4V2.5C5.5 2.22386 5.72386 2 6 2H10C10.2761 2 10.5 2.22386 10.5 2.5V4M6.5 7V11.5M9.5 7V11.5M3.5 4L4.25 13.5C4.25 13.7761 4.47386 14 4.75 14H11.25C11.5261 14 11.75 13.7761 11.75 13.5L12.5 4"
-                                                        stroke="currentColor"
-                                                        strokeWidth="1.5"
-                                                        strokeLinecap="round"
-                                                    />
-                                                </svg>
-                                                <p className="minecraft-seven">Delete Profile</p>
-                                            </div>
-                                        </div>,
-                                        document.body
-                                    )}
-                            </div>
+                            <ThreeDotsMenu items={[
+                                {
+                                    label: "Open Folder",
+                                    icon: <svg width="14" height="14" viewBox="0 0 16 16" fill="none"><path d="M1 3C1 2.44772 1.44772 2 2 2H6.17157C6.43679 2 6.69114 2.10536 6.87868 2.29289L7.70711 3.12132C7.89464 3.30886 8.149 3.41421 8.41421 3.41421H14C14.5523 3.41421 15 3.86193 15 4.41421V13C15 13.5523 14.5523 14 14 14H2C1.44772 14 1 13.5523 1 13V3Z" stroke="#FFFFFF" strokeWidth="1.5" /></svg>,
+                                    onClick: openProfileFolder,
+                                },
+                                {
+                                    label: "Delete Profile",
+                                    icon: <svg width="14" height="14" viewBox="0 0 16 16" fill="none"><path d="M2 4H14M5.5 4V2.5C5.5 2.22386 5.72386 2 6 2H10C10.2761 2 10.5 2.22386 10.5 2.5V4M6.5 7V11.5M9.5 7V11.5M3.5 4L4.25 13.5C4.25 13.7761 4.47386 14 4.75 14H11.25C11.5261 14 11.75 13.7761 11.75 13.5L12.5 4" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" /></svg>,
+                                    onClick: deleteProfile,
+                                    danger: true,
+                                },
+                            ]} />
                         </div>
                         <MinecraftButton
                             text="Add Content"
